Move test to compile task (compile.all is an impl detail)
